Move qualifiedTitle to Configuration, redo exports
